Move filesystem functions to `base/fs.cpp/h`
